requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#

cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.21.0  # via google-cloud-bigquery, google-cloud-core, google-cloud-datastore
google-auth==1.18.0       # via google-api-core, google-cloud-bigquery
google-cloud-bigquery==1.25.0  # via -r requirements.in
google-cloud-core==1.3.0  # via google-cloud-bigquery, google-cloud-datastore
google-cloud-datastore==1.12.0  # via -r requirements.in
google-resumable-media==0.5.1  # via google-cloud-bigquery
googleapis-common-protos==1.52.0  # via google-api-core
grpcio==1.30.0            # via google-api-core
idna==2.10                # via requests
protobuf==3.12.2          # via google-api-core, google-cloud-bigquery, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pytz==2020.1              # via google-api-core
requests==2.24.0          # via google-api-core
rsa==4.6                  # via google-auth
six==1.15.0               # via google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, grpcio, protobuf
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
